Work with PP services with subdirectory URLs
In case the PhotoPrism is served from subdirectory (e.g.
`https://domain.com/photoprism/`) a bit broader match rule is
required.

In theory, the more allowing rule might execute the script in URLs
other than PhotoPrism. However, in practice the script itself creates
the buttons only if correctly named elements are found in the page.

This drawback can be effectively ruled out by making the extension
permissions tighter than the default. This is now documented in the
README.md.